+++ common/src/main/java/de/maxhenkel/voicechat/gui/audiodevice/SelectDeviceScreen.java
@@ -1,5 +1,6 @@
 package de.maxhenkel.voicechat.gui.audiodevice;
 
+import de.maxhenkel.configbuilder.entry.ConfigEntry;
 import de.maxhenkel.voicechat.Voicechat;
 import de.maxhenkel.voicechat.gui.VoiceChatScreenBase;
 import de.maxhenkel.voicechat.gui.widgets.ListScreenBase;
@@ -7,10 +8,8 @@ import net.minecraft.ChatFormatting;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.components.Button;
 import net.minecraft.client.gui.screens.Screen;
-import net.minecraft.client.resources.sounds.SimpleSoundInstance;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.sounds.SoundEvents;
 import net.minecraft.util.Mth;
 
 import javax.annotation.Nullable;
@@ -24,7 +23,7 @@ public abstract class SelectDeviceScreen extends ListScreenBase {
     protected static final int HEADER_SIZE = 16;
     protected static final int FOOTER_SIZE = 32;
     protected static final int UNIT_SIZE = 18;
-    protected static final int CELL_HEIGHT = 36;
+
 
     @Nullable
     protected Screen parent;
@@ -39,29 +38,25 @@ public abstract class SelectDeviceScreen extends ListScreenBase {
 
     public abstract List<String> getDevices();
 
-    public abstract String getSelectedDevice();
-
-    public abstract void onSelect(String device);
-
-    public abstract ResourceLocation getIcon(String device);
+    public abstract ResourceLocation getIcon();
 
     public abstract Component getEmptyListComponent();
 
-    public abstract String getVisibleName(String device);
+    public abstract ConfigEntry<String> getConfigEntry();
 
     @Override
     protected void init() {
         super.init();
         guiLeft = guiLeft + 2;
         guiTop = 32;
-        int minUnits = Mth.ceil((float) (CELL_HEIGHT + 4) / (float) UNIT_SIZE);
+        int minUnits = Mth.ceil((float) (AudioDeviceList.CELL_HEIGHT + 4) / (float) UNIT_SIZE);
         units = Math.max(minUnits, (height - HEADER_SIZE - FOOTER_SIZE - guiTop * 2) / UNIT_SIZE);
         ySize = HEADER_SIZE + units * UNIT_SIZE + FOOTER_SIZE;
 
         if (deviceList != null) {
             deviceList.setRectangle(width, units * UNIT_SIZE, 0, guiTop + HEADER_SIZE);
         } else {
-            deviceList = new AudioDeviceList(width, units * UNIT_SIZE, guiTop + HEADER_SIZE, CELL_HEIGHT);
+            deviceList = new AudioDeviceList(width, units * UNIT_SIZE, guiTop + HEADER_SIZE).setIcon(getIcon()).setConfigEntry(getConfigEntry());
         }
         addWidget(deviceList);
 
@@ -70,7 +65,7 @@ public abstract class SelectDeviceScreen extends ListScreenBase {
         }).bounds(guiLeft + 7, guiTop + ySize - 20 - 7, xSize - 14, 20).build();
         addRenderableWidget(back);
 
-        deviceList.replaceEntries(getDevices().stream().map(s -> new AudioDeviceEntry(this, s)).toList());
+        deviceList.setAudioDevices(getDevices());
     }
 
 
@@ -94,23 +89,6 @@ public abstract class SelectDeviceScreen extends ListScreenBase {
         } else {
             guiGraphics.drawCenteredString(font, getEmptyListComponent(), width / 2, guiTop + HEADER_SIZE + (units * UNIT_SIZE) / 2 - font.lineHeight / 2, -1);
         }
-    }
-
-    @Override
-    public boolean mouseClicked(double mouseX, double mouseY, int button) {
-        if (super.mouseClicked(mouseX, mouseY, button)) {
-            return true;
-        }
-        for (AudioDeviceEntry entry : deviceList.children()) {
-            if (entry.isMouseOver(mouseX, mouseY)) {
-                if (!getSelectedDevice().equals(entry.getDevice())) {
-                    minecraft.getSoundManager().play(SimpleSoundInstance.forUI(SoundEvents.UI_BUTTON_CLICK, 1F));
-                    onSelect(entry.getDevice());
-                    return true;
-                }
-            }
-        }
-        return false;
     }
 
 }