+++ common/src/main/java/de/maxhenkel/voicechat/gui/widgets/MicTestButton.java
@@ -8,6 +8,7 @@ import de.maxhenkel.voicechat.voice.client.speaker.Speaker;
 import de.maxhenkel.voicechat.voice.client.speaker.SpeakerException;
 import de.maxhenkel.voicechat.voice.client.speaker.SpeakerManager;
 import de.maxhenkel.voicechat.voice.common.Utils;
+import net.minecraft.ChatFormatting;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.components.AbstractButton;
 import net.minecraft.client.gui.narration.NarrationElementOutput;
@@ -18,9 +19,8 @@ import java.util.function.Consumer;
 
 public class MicTestButton extends AbstractButton {
 
-    private static final Component TEST_UNAVAILABLE = Component.translatable("message.voicechat.mic_test_unavailable");
-    private static final Component TEST_ON = Component.translatable("message.voicechat.mic_test_on");
-    private static final Component TEST_OFF = Component.translatable("message.voicechat.mic_test_off");
+    private static final Component TEST = Component.translatable("message.voicechat.mic_test");
+    private static final Component TEST_UNAVAILABLE = Component.translatable("message.voicechat.mic_test_unavailable").withStyle(ChatFormatting.RED);
 
     private boolean micActive;
     @Nullable
@@ -30,23 +30,10 @@ public class MicTestButton extends AbstractButton {
     private final ClientVoicechat client;
 
     public MicTestButton(int xIn, int yIn, int widthIn, int heightIn, MicListener micListener) {
-        super(xIn, yIn, widthIn, heightIn, Component.empty());
+        super(xIn, yIn, widthIn, heightIn, TEST);
         this.micListener = micListener;
         this.client = ClientManager.getClient();
         active = client == null || client.getSoundManager() != null;
-        updateText();
-    }
-
-    private void updateText() {
-        if (!active) {
-            setMessage(TEST_UNAVAILABLE);
-            return;
-        }
-        if (micActive) {
-            setMessage(TEST_ON);
-        } else {
-            setMessage(TEST_OFF);
-        }
     }
 
     @Override
@@ -59,7 +46,14 @@ public class MicTestButton extends AbstractButton {
 
     public void setMicActive(boolean micActive) {
         this.micActive = micActive;
-        updateText();
+    }
+
+    @Nullable
+    public Component getHoverText() {
+        if (!active) {
+            return TEST_UNAVAILABLE;
+        }
+        return null;
     }
 
     @Override
@@ -82,7 +76,6 @@ public class MicTestButton extends AbstractButton {
         } else {
             close();
         }
-        updateText();
     }
 
     private void close() {
@@ -166,6 +159,7 @@ public class MicTestButton extends AbstractButton {
             if (ownSoundManager != null) {
                 ownSoundManager.close();
             }
+            micActive = false;
             Voicechat.LOGGER.info("Mic test audio channel closed");
         }