+++ neoforge/src/main/java/de/maxhenkel/voicechat/NeoForgeVoicechatClientMod.java
@@ -1,6 +1,7 @@
 package de.maxhenkel.voicechat;
 
 import de.maxhenkel.voicechat.gui.VoiceChatSettingsScreen;
+import de.maxhenkel.voicechat.gui.onboarding.OnboardingManager;
 import de.maxhenkel.voicechat.integration.clothconfig.ClothConfig;
 import de.maxhenkel.voicechat.integration.clothconfig.ClothConfigIntegration;
 import de.maxhenkel.voicechat.intercompatibility.ClientCompatibilityManager;
@@ -23,6 +24,9 @@ public class NeoForgeVoicechatClientMod extends VoicechatClient {
         NeoForge.EVENT_BUS.register(ClientCompatibilityManager.INSTANCE);
         ClothConfig.init();
         ModLoadingContext.get().registerExtensionPoint(ConfigScreenHandler.ConfigScreenFactory.class, () -> new ConfigScreenHandler.ConfigScreenFactory((client, parent) -> {
+            if (OnboardingManager.isOnboarding()) {
+                return OnboardingManager.getOnboardingScreen(parent);
+            }
             if (ClothConfig.isLoaded()) {
                 return ClothConfigIntegration.createConfigScreen(parent);
             } else {